requirements.txt
# Core NLP and ML
spacy>=3.7.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0

# LangChain & AI
langchain>=0.1.0
langchain-groq>=0.0.1
langchain-community>=0.0.20
groq>=0.4.0
sentence-transformers>=2.2.2

# Text processing
nltk>=3.8.0
pdfplumber>=0.10.0
python-docx>=1.0.0

# UML Generation
plantuml>=0.3.0
requests>=2.31.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Utilities
python-multipart>=0.0.6
jinja2>=3.1.2
python-dotenv>=1.0.0
aiofiles>=23.2.1
